+++ .gitignore
@@ -1,1 +1,2 @@
 /gunk
+/google
