+++ .travis.yml
@@ -1,23 +1,21 @@
 language: go
-
 go:
   - 1.9.x
-
+  - 1.10.x
+  - tip
 sudo: false
-
-cache:
-  directories:
-  - $HOME/protobuf-3.5.1
-
-env:
-  global:
-  - PATH=$HOME/protobuf-3.5.1/bin:$PATH
-
+addons:
+  apt:
+    packages:
+      - unzip
 before_install:
-  - ./install-protobuf.sh
-
+  - mkdir -p $PWD/protoc /tmp/protobuf
+  - wget -O $PWD/protoc/protoc https://repo1.maven.org/maven2/com/google/protobuf/protoc/3.5.1-1/protoc-3.5.1-1-linux-x86_64.exe
+  - wget -O /tmp/protobuf/protobuf.jar https://repo1.maven.org/maven2/com/google/protobuf/protobuf-java/3.5.1/protobuf-java-3.5.1.jar
+  - unzip -qq /tmp/protobuf/protobuf.jar -d /tmp/protobuf && mv /tmp/protobuf/google $PWD
+  - chmod +x $PWD/protoc/protoc
+  - export PATH=$PWD/protoc/:$PATH
 install: true
-
 script:
   - go install ./vendor/github.com/golang/protobuf/protoc-gen-go
   # use -short to skip the "go build" tests
